COG-281: MessagesCodes is renamed to MessagesCode
